chore: improve ButtonGroup exports
